+++ packages/control-hub/src/index.ts
@@ -1,15 +1,2 @@
-import { ControlHubInternal } from "./internal/ControlHub.js";
-import { ControlHub } from "@rev-robotics/rev-hub-core";
-
-export { openConnectedControlHub } from "./discovery.js";
-export { openUsbControlHubsAndChildren } from "./internal/ControlHub.js";
-
-export async function openControlHub(
-    serialNumber: string,
-    moduleAddress: number,
-    port: number,
-): Promise<ControlHub> {
-    let hub = new ControlHubInternal(serialNumber);
-    await hub.open("127.0.0.1", (port + 1).toString());
-    return hub;
-}
+export * from "./open-control-hub.js";
+export * from "./discovery.js";
