+++ packages/core/src/index.ts
@@ -22,6 +22,7 @@ export * from "./SerialParity.js";
 export * from "./VerbosityLevel.js";
 export * from "./Version.js";
 export * from "./serial-errors.js";
+export * from "./nack-errors/set-prototype.js";
 export * from "./nack-errors/nack-codes.js";
 export * from "./nack-errors/NackError.js";
 export * from "./nack-errors/BatteryTooLowError.js";
