+++ packages/core/src/nack-errors/set-prototype.ts
@@ -1,0 +1,4 @@
+export function setPrototypeOf(obj: any, proto: any) {
+    Object.setPrototypeOf(obj, proto);
+    return obj;
+}
