+++ packages/expansion-hub/src/open-rev-hub.ts
@@ -22,13 +22,7 @@ import { SerialPort as SerialLister } from "serialport";
 import { ExpansionHubInternal } from "./internal/ExpansionHub.js";
 import { startKeepAlive } from "./start-keep-alive.js";
 import { performance } from "perf_hooks";
-
-/**
- * Maps the serial port path (/dev/tty1 or COM3 for example) to an open
- * Serial object at that path. The {@link SerialPort} object should be removed from
- * the map upon closing.
- */
-const openSerialMap = new Map<string, typeof NativeSerial>();
+import { getSerial } from "./serial.js";
 
 /**
  *
@@ -65,11 +59,7 @@ export async function openParentExpansionHub(
 ): Promise<ParentExpansionHub> {
     let serialPortPath = await getSerialPortPathForExHubSerial(serialNumber);
 
-    if (openSerialMap.get(serialPortPath) == undefined) {
-        openSerialMap.set(serialPortPath, await openSerialPort(serialPortPath));
-    }
-
-    let serialPort = openSerialMap.get(serialPortPath)!;
+    let serialPort = await getSerial(serialPortPath);
 
     let parentHub = new ExpansionHubInternal(true, serialPort, serialNumber);
 
@@ -121,11 +111,7 @@ export async function openExpansionHubAndAllChildren(
 ): Promise<ParentExpansionHub> {
     let serialPortPath = await getSerialPortPathForExHubSerial(serialNumber);
 
-    if (openSerialMap.get(serialPortPath) == undefined) {
-        openSerialMap.set(serialPortPath, await openSerialPort(serialPortPath));
-    }
-
-    let serialPort = openSerialMap.get(serialPortPath)!;
+    let serialPort = await getSerial(serialPortPath);
 
     let discoveredModules = await NativeRevHub.discoverRevHubs(serialPort);
     let parentAddress = discoveredModules.parentAddress;
@@ -153,47 +139,4 @@ async function getSerialPortPathForExHubSerial(serialNumber: string): Promise<st
         }
     }
     throw new Error(`Unable to find serial port for ${serialNumber}`);
-}
-
-/**
- * Closes the given Serial port and removes it from the open serial ports
- * list. This should be the preferred way to close a Serial port.
- *
- * @param serialPort the Serial port to close
- */
-export function closeSerialPort(serialPort: typeof NativeSerial) {
-    for (let [path, port] of openSerialMap.entries()) {
-        if (port === serialPort) {
-            openSerialMap.delete(path);
-        }
-    }
-    serialPort.close();
-}
-
-async function openSerialPort(serialPortPath: string): Promise<typeof NativeSerial> {
-    let serial = new NativeSerial();
-    try {
-        await serial.open(
-            serialPortPath,
-            460800,
-            8,
-            SerialParity.None,
-            1,
-            SerialFlowControl.None,
-        );
-    } catch (e: any) {
-        let code = e.errorCode;
-        if (code == SerialError.INVALID_ARGS) {
-            throw new InvalidSerialArguments(serialPortPath);
-        } else if (code == SerialError.UNABLE_TO_OPEN) {
-            throw new UnableToOpenSerialError(serialPortPath);
-        } else if (code == SerialError.CONFIGURATION_ERROR) {
-            throw new SerialConfigurationError(serialPortPath);
-        } else if (code == SerialError.IO_ERROR) {
-            throw new SerialIoError(serialPortPath);
-        } else if (code == SerialError.GENERAL_ERROR) {
-            throw new GeneralSerialError(serialPortPath);
-        }
-    }
-    return serial;
 }
