+++ packages/sample/package.json
@@ -3,7 +3,7 @@
     "descripton": "sample for testing Rev Hub code",
     "version": "1.0.0",
     "main": "dist/main.js",
-    "bin": "dist/main.js",
+    "bin": "revhub",
     "private": true,
     "type": "module",
     "dependencies": {
@@ -17,6 +17,12 @@
     "devDependencies": {
         "@types/ws": "^8.5.4",
         "@types/node-forge": "^1.3.2",
+        "@rollup/plugin-node-resolve": "^15.1.0",
+        "@rollup/plugin-typescript": "^11.1.1",
+        "@rollup/plugin-commonjs": "^25.0.1",
+        "rollup-plugin-exclude-dependencies-from-bundle": "^1.1.23",
+        "@rollup/plugin-json": "^6.0.0",
+        "rollup-plugin-natives": "^0.7.6",
         "@types/debug": "^4.1.8"
     },
     "scripts": {
